+++ src/aux-server/aux-web/aux-projector/BuilderApp/BuilderApp.ts
@@ -393,7 +393,7 @@ export default class BuilderApp extends Vue {
                                             this.showCreateChannel = true;
                                             this.snackbar = {
                                                 message:
-                                                    'This channel does not exist. Do you want to create it?',
+                                                    'Channel does not exist. Tap here to create it.',
                                                 visible: true,
                                                 action: {
                                                     label: 'Create Channel',
@@ -681,6 +681,10 @@ export default class BuilderApp extends Vue {
         });
 
         await appManager.simulationManager.removeSimulation('admin');
+
+        setTimeout(() => {
+            this.refreshPage();
+        }, 500);
     }
 
     refreshPage() {